Created a exam insertion view, also now TodayView can display exams inserted previously.

+++ app/src/main/java/net/felipealafy/studentplanner/tablemodels/ExamTable.kt
@@ -22,7 +22,7 @@ data class ExamTable(
     val subjectId: String,
     val name: String,
     val grade: Float,
-    val gradeWeight: Int,
+    val gradeWeight: Float,
     val start: LocalDateTime,
     val end: LocalDateTime
 )
